Moved KeyManager out of config package
